Revert "[Rename] 패키지 위치 변경"
This reverts commit ece313f9421169910a98eb544f4416a0208b183a.
