+++ src/main/java/oop/blackjack/v2/ApplicationV2.java
@@ -1,4 +1,4 @@
-package oop.blackjackv2;
+package oop.blackjack.v2;
 
 public class ApplicationV2 {
 }
