+++ src/main/java/oop/blackjack/v2/domain/Rule.java
@@ -1,0 +1,4 @@
+package oop.blackjack.v2.domain;
+
+public class Rule {
+}
