--- src/main/java/oop/blackjackv2/domain/CardDeck.java
@@ -1,4 +1,0 @@
-package oop.blackjackv2.domain;
-
-public class CardDeck {
-}
