--- src/main/java/oop/blackjackv2/domain/RealGame.java
@@ -1,4 +1,0 @@
-package oop.blackjackv2.domain;
-
-public class RealGame {
-}
