--- src/main/java/oop/blackjackv2/domain/View.java
@@ -1,4 +1,0 @@
-package oop.blackjackv2.domain;
-
-public class View {
-}
